Fixed the Logoff buttons and more basic things that were deleted from the last push.
